# Return error stack traces for the CLI client