@wadells 2021-09-01
Sorry I didn't see this (and thus provide feedback) earlier.

Did you consider a server side debug flag?  Scanning software won't have this header set, but it is ultimately up to the client whether these are exposed -- which feels like it weakens the hardening done in https://github.com/gravitational/teleport/pull/5070.