@a-palchikov 2021-09-02
> Sorry I didn't see this (and thus provide feedback) earlier.
> 
> Did you consider gating this behind a server side debug flag? Scanning software won't have this header set, but with the current implementation it is solely up to the client whether these are exposed -- which feels like it weakens the hardening done in [gravitational/teleport#5070](https://github.com/gravitational/teleport/pull/5070).

I have not thought about a debug flag but it seems to only further complicate things. Besides it's not going to help debug an issue when the gravity client is talking to an existing infra which might be difficult to restart with a debug flag.
I agree this is not the ideal solution to eliminate exposure but so far the troubleshooting benefits outweigh anything slightly more complex I've thought about.
The header value can be generated dynamically to make the guessing more challenging but I don't think it's worth it.